Flask-MongoDB-recipe-book/requirements.txt
bson==0.5.8
Click==7.0
dnspython==1.16.0
Flask==1.0.2
Flask-PyMongo==2.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
passlib==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
Werkzeug==0.15.2
WTForms==2.2.1
